chore: split dal code into multiple files
