+++ cmd/launch_screen/dal/db/add_image_list_showtime.go
@@ -1,0 +1,16 @@
+package db
+
+import (
+	"context"
+	"fmt"
+)
+
+func AddImageListShowTime(ctx context.Context, pictureList *[]Picture) error {
+	for i := range *pictureList {
+		(*pictureList)[i].ShowTimes++
+	}
+	if err := DB.WithContext(ctx).Save(pictureList).Error; err != nil {
+		return fmt.Errorf("dal.AddImageListShowTime error: %v", err)
+	}
+	return nil
+}
