+++ cmd/paper/dal/cache/util.go
@@ -1,0 +1,36 @@
+/*
+Copyright 2024 The west2-online Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cache
+
+import (
+	"context"
+	"strings"
+
+	"github.com/west2-online/fzuhelper-server/pkg/constants"
+)
+
+func IsExists(ctx context.Context, key string) bool {
+	return RedisClient.Exists(ctx, key).Val() == 1
+}
+
+func getFileDirKey(path string) string {
+	keys := []string{
+		constants.CACHE_FILEDIR,
+		path,
+	}
+	return strings.Join(keys, "_")
+}
